feat: update parameter consume form design
