+++ src/components/Asset/AssetActions/Compute/ConsumerParameters/index.module.css
@@ -1,0 +1,8 @@
+.container {
+  margin-top: -1rem;
+  margin-bottom: 1rem;
+  margin-left: -2rem;
+  margin-right: -2rem;
+  border-top: 1px solid var(--border-color);
+  border-bottom: 1px solid var(--border-color);
+}
