+++ src/components/Asset/AssetActions/Compute/FormComputeDataset.module.css
@@ -26,12 +26,3 @@
 .warning {
   margin-bottom: var(--spacer);
 }
-
-.consumerParamsContainer {
-  margin-top: -1rem;
-  margin-bottom: 1rem;
-  margin-left: -2rem;
-  margin-right: -2rem;
-  border-top: 1px solid var(--border-color);
-  border-bottom: 1px solid var(--border-color);
-}
