+++ src/components/Asset/AssetActions/Compute/FormComputeDataset.tsx
@@ -17,8 +17,6 @@ import { getTokenBalanceFromSymbol } from '@utils/web3'
 import { MAX_DECIMALS } from '@utils/constants'
 import Decimal from 'decimal.js'
 import ConsumerParameters from './ConsumerParameters'
-import Tabs from '@components/@shared/atoms/Tabs'
-import { parseConsumerParameters } from '@utils/ddo'
 
 export default function FormStartCompute({
   algorithms,
@@ -235,76 +233,6 @@ export default function FormStartCompute({
     })
   }, [balance, dtBalance, datasetSymbol, algorithmSymbol, totalPrices])
 
-  // const updateTabs = useCallback(() => {
-  //   const tabs = []
-  //   if (selectedAlgorithmAsset?.services[0]?.consumerParameters) {
-  //     tabs.push({
-  //       title: 'Algo service',
-  //       content: (
-  //         <ConsumerParameters
-  //           parameters={selectedAlgorithmAsset.services[0]?.consumerParameters}
-  //         />
-  //       )
-  //     })
-  //   }
-  //   if (selectedAlgorithmAsset?.metadata?.algorithm?.consumerParameters) {
-  //     tabs.push({
-  //       title: 'Algo service',
-  //       content: (
-  //         <ConsumerParameters
-  //           parameters={
-  //             selectedAlgorithmAsset.metadata?.algorithm?.consumerParameters
-  //           }
-  //         />
-  //       )
-  //     })
-  //   }
-  //   return tabs
-  // }, [selectedAlgorithmAsset])
-  // const [tabs, setTabs] = useState(updateTabs())
-
-  // useEffect(() => {
-  //   setTabs(updateTabs())
-  // }, [updateTabs])
-
-  const tabs = []
-  if (asset?.services[0]?.consumerParameters) {
-    tabs.push({
-      title: 'Data Service',
-      content: (
-        <ConsumerParameters
-          parameters={parseConsumerParameters(
-            asset.services[0]?.consumerParameters
-          )}
-        />
-      )
-    })
-  }
-  if (selectedAlgorithmAsset?.services[0]?.consumerParameters) {
-    tabs.push({
-      title: 'Algo Service',
-      content: (
-        <ConsumerParameters
-          parameters={parseConsumerParameters(
-            selectedAlgorithmAsset.services[0]?.consumerParameters
-          )}
-        />
-      )
-    })
-  }
-  if (selectedAlgorithmAsset?.metadata?.algorithm?.consumerParameters) {
-    tabs.push({
-      title: 'Algo Params',
-      content: (
-        <ConsumerParameters
-          parameters={parseConsumerParameters(
-            selectedAlgorithmAsset.metadata?.algorithm?.consumerParameters
-          )}
-        />
-      )
-    })
-  }
-
   return (
     <Form className={styles.form}>
       <Alert
@@ -323,10 +251,11 @@ export default function FormStartCompute({
           />
         )
       })}
-      {selectedAlgorithmAsset && (
-        <div className={styles.consumerParamsContainer}>
-          <Tabs items={tabs} />
-        </div>
+      {asset && selectedAlgorithmAsset && (
+        <ConsumerParameters
+          asset={asset}
+          selectedAlgorithmAsset={selectedAlgorithmAsset}
+        />
       )}
       <PriceOutput
         hasPreviousOrder={hasPreviousOrder}
